refactor: Move BasicPathTestCase into transferoptimization where it is used.
